added fn options sim,nao and modules concept
