+++ src/lib.rs
@@ -1,0 +1,3 @@
+pub mod terminal;
+pub mod todo;
+pub mod terminalerror;
